Fixed possible decoding errors (the duration can be .25 seconds?!)

+++ Bot/src/main/java/com/evilco/bot/PlugBot/core/data/PlugMedia.java
@@ -20,7 +20,7 @@ public class PlugMedia {
 	/**
 	 * The media duration in seconds.
 	 */
-	public int duration;
+	public double duration;
 
 	/**
 	 * ??? (1 = YouTube, 2/0 = SoundCloud)
